third_party/blink/web_tests/virtual/shared_array_buffer_on_desktop/fast/workers/worker-sharedarraybuffer-transfer-expected.txt

Test sharing SharedArrayBuffers between dedicated workers.

Running SharedArrayBuffer test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred data is OK.
Running Int8Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint8Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint8ClampedArray test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Int16Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint16Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Int32Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Uint32Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Float32Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running Float64Array test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: SharedArrayBuffer not neutered during transfer.
PASS: Original data not changed during transfer.
PASS: Transferred array type is OK.
PASS: Transferred data is OK.
Running SendSharedArrayBufferTwice test case
PASS: Transferred data is OK.
Running TransferArrayBufferAndSharedArrayBuffer test case
PASS: ArrayBuffer was neutered during transfer.
PASS: Transferred data is OK.
PASS: Transferred data is OK.
Running SharedArrayBufferInTransferListViaMessagePort test case
PASS: Passing SharedArrayBuffer in the transfer list threw.
PASS: ArrayBuffer 'ab1' was not neutered.
PASS: ArrayBuffer 'ab2' was not neutered.
DONE
